Move idbatcher to processor/tailsampling
