+++ processor/tailsampling/processor_test.go
@@ -25,8 +25,8 @@ import (
 
 	"github.com/open-telemetry/opentelemetry-service/consumer/consumerdata"
 	"github.com/open-telemetry/opentelemetry-service/exporter/exportertest"
-	"github.com/open-telemetry/opentelemetry-service/internal/collector/processor/idbatcher"
 	"github.com/open-telemetry/opentelemetry-service/processor"
+	"github.com/open-telemetry/opentelemetry-service/processor/tailsampling/idbatcher"
 	"github.com/open-telemetry/opentelemetry-service/processor/tailsampling/sampling"
 	tracetranslator "github.com/open-telemetry/opentelemetry-service/translator/trace"
 )
